// helpers/helpers.tsx
import { ProductIcon } from "@/icon-components/ProductIcon";
import { FirstLevelMenuItem } from "@/interfaces/menu.interface";
import { CoursesIcon } from "@/icon-components/CoursesIcon";
import { TopLevelCategory } from "@/interfaces/page.interface";
import { ServicesIcon } from "@/icon-components/ServicesIcon";
import { BookIcon } from "@/icon-components/BookIcon";

export const firstLevelMenu: FirstLevelMenuItem[] = [
  {
    route: "[type]",
    name: "Курсы",
    icon: <CoursesIcon />,
    id: TopLevelCategory.Courses,
  },
  {
    route: "services",
    name: "Сервисы",
    icon: <ServicesIcon />,
    id: TopLevelCategory.Services,
  },
  {
    route: "books",
    name: "Книги",
    icon: <BookIcon />,
    id: TopLevelCategory.Books,
  },
  {
    route: "products",
    name: "Продукты",
    icon: <ProductIcon />,
    id: TopLevelCategory.Products,
  },
];

export const priceRu = (price: number): string =>
  price
    .toString()
    .replace(/\B(?=(\d{3})+(?!\d))/g, " ")
    .concat(" ₽");

export const declOfNum = (
  number: number,
  titles: [string, string, string],
): string => {
  const cases = [2, 0, 1, 1, 1, 2];
  return titles[
    number % 100 > 4 && number % 100 < 20
      ? 2
      : cases[number % 10 < 5 ? number % 10 : 5]
  ];
};
